chore(upload->demo): split file && add previewColumns fn demo
